fix:user: store birthdate instead of date
This is usually better to manipulate a date for age, from there we can
derive the age.

I rewrite this because the age field prevents me to create a superuser
to test on my side, because it had a non-null constraints which was
incompatible with the django cli.

When the user register we can omit its birthdate for now, later if the
user want to enrich its profile, it can always set it up there.

Signed-off-by: Lionel H <me@nullbyte.be>