setup github actions
ci.yml hold the ci related jobs. Everytime a new commit is pushed
modifying a python file or the action files themselves, the job will be
triggered and running tests.

issue-on-failure.yml is trigged when a workflow failed on the develop
branch, for example when we merged a PR on develop. Since develop is the
protected default branch, builds on this branch should have been
validated and should never fail ! If that would be the case, an issue
would be automatically created for us to take action because it
basically means that the project is broken.

Signed-off-by: Lionel H <me@nullbyte.be>